services/detection/requirements/base.txt
# Base
Cython
opencv-python>=4.1.2
Pillow
PyYAML>=5.3.1
tqdm>=4.41.0
imageio
numpy-stl

# Machine Learning
shapely==1.8.1
timm==0.6.11
mmdet==2.28.1
mmengine

# Logging
wandb
nvitop
colorlog

# Plotting
seaborn>=0.11.0
pandas

# Utilities
python-dotenv==0.19.0
requests==2.26.0
boto3==1.35.94
botocore==1.35.94
s3transfer==0.10.0
termcolor
yacs
scikit-learn
yapf==0.40.1
